Ensure tests don't fail erroneously
Hypothesis generated a version number like `-v`, which triggered
a different flag in the script arguments.

Add a lint-only mode to the test script. If you don't have to wait
for the tests to run it is easier to make changes for linting.

+++ test.sh
@@ -2,17 +2,20 @@
 
 set -euo pipefail
 
-image_id=cdflow-wrapper.test
+image_id="$(basename $(pwd))"
 
 docker build -t "${image_id}" .
 
-docker run --rm --name "${image_id}" \
-    -i $(tty -s && echo -t) \
-    "${image_id}" \
-    py.test \
-    --cov=. \
-    --cov-report term-missing \
-    "$@"
+if [ ${1} != 'lint-only' ]
+then
+    docker run --rm --name "${image_id}" \
+        -i $(tty -s && echo -t) \
+        "${image_id}" \
+        py.test \
+        --cov=. \
+        --cov-report term-missing \
+        "${@}"
+fi
 
 docker run --rm --name "${image_id}" "${image_id}" \
     flake8 --max-complexity=5
